Refactor XPath into module
It seems much more like the other modules like random than a core part of running an applet.
